--- a/extensions/micrometer/runtime/src/main/java/io/quarkus/micrometer/runtime/MicrometerCountedInterceptor.java
+++ b/extensions/micrometer/runtime/src/main/java/io/quarkus/micrometer/runtime/MicrometerCountedInterceptor.java
@@ -1,6 +1,5 @@
 package io.quarkus.micrometer.runtime;
 
-import java.lang.reflect.Method;
 import java.util.concurrent.CompletionStage;
 import java.util.function.BiConsumer;
 
@@ -13,6 +12,8 @@ import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tags;
 import io.quarkus.arc.ArcInvocationContext;
+import io.quarkus.arc.MethodMetadata;
+import io.quarkus.arc.Reflectionless;
 import io.smallrye.mutiny.Uni;
 import io.smallrye.mutiny.tuples.Functions;
 
@@ -24,6 +25,7 @@ import io.smallrye.mutiny.tuples.Functions;
  * @see Counted
  */
 @Interceptor
+@Reflectionless
 @MicrometerCounted
 @Priority(Interceptor.Priority.LIBRARY_BEFORE + 10)
 public class MicrometerCountedInterceptor {
@@ -60,7 +62,7 @@ public class MicrometerCountedInterceptor {
         if (counted == null) {
             return context.proceed();
         }
-        Method method = context.getMethod();
+        MethodMetadata method = context.getMethodMetadata();
         Tags commonTags = getCommonTags(method.getDeclaringClass().getName(), method.getName());
 
         Class<?> returnType = method.getReturnType();