--- a/independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/BeanInfo.java
+++ b/independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/BeanInfo.java
@@ -835,6 +835,14 @@ public class BeanInfo implements InjectionTargetInfo {
             return interceptors.isEmpty();
         }
 
+        boolean isReflectionless() {
+            for (InterceptorInfo interceptor : interceptors) {
+                if (!interceptor.isReflectionless()) {
+                    return false;
+                }
+            }
+            return true;
+        }
     }
 
     static class DecorationInfo {