--- a/independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/DecoratorGenerator.java
+++ b/independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/DecoratorGenerator.java
@@ -43,11 +43,13 @@ public class DecoratorGenerator extends BeanGenerator {
     protected static final String FIELD_NAME_DELEGATE_TYPE = "delegateType";
     static final String ABSTRACT_IMPL_SUFFIX = "_Impl";
 
-    public DecoratorGenerator(AnnotationLiteralProcessor annotationLiterals, Predicate<DotName> applicationClassPredicate,
+    public DecoratorGenerator(AnnotationLiteralProcessor annotationLiterals, MethodMetadataProcessor methodsMetadata,
+            Predicate<DotName> applicationClassPredicate,
             PrivateMembersCollector privateMembers, boolean generateSources, ReflectionRegistration reflectionRegistration,
             Set<String> existingClasses, Map<BeanInfo, String> beanToGeneratedName,
             Predicate<DotName> injectionPointAnnotationsPredicate) {
-        super(annotationLiterals, applicationClassPredicate, privateMembers, generateSources, reflectionRegistration,
+        super(annotationLiterals, methodsMetadata, applicationClassPredicate, privateMembers, generateSources,
+                reflectionRegistration,
                 existingClasses, beanToGeneratedName, injectionPointAnnotationsPredicate, Collections.emptyList());
     }