… extensions/arc/deployment/src/test/java/io/quarkus/arc/test/interceptor/staticmethods/InterceptedStaticMethodReflectionlessTest.java
… independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/InterceptorInfo.java
… independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/MethodDescriptors.java
… independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/MethodMetadataGenerator.java
… independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/MethodMetadataProcessor.java
… independent-projects/arc/processor/src/main/java/io/quarkus/arc/processor/SubclassGenerator.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/ArcInvocationContext.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/MethodMetadata.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/ParameterMetadata.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/Reflectionless.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/AbstractInvocationContext.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/AnnotationLiterals.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/AroundConstructInvocationContext.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/AroundInvokeInvocationContext.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/InterceptedMethodMetadata.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/InterceptedStaticMethods.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/InvocationContexts.java
… independent-projects/arc/runtime/src/main/java/io/quarkus/arc/impl/LifecycleCallbackInvocationContext.java
… independent-projects/arc/tests/src/test/java/io/quarkus/arc/test/interceptors/reflectionless/AroundConstructReflectionlessInterceptorTest.java
… independent-projects/arc/tests/src/test/java/io/quarkus/arc/test/interceptors/reflectionless/AroundInvokeReflectionlessInterceptorTest.java