This is a result of my attempt to optimize reflection usage in interceptors by providing an reflectionless alternative to `InvocationContext.getMethod()` / `getConstructor()`. This has 2 parts:

1. The `ArcInvocationContext` gains a new method `getMethodMetadata()`, which provides an alternative to `getMethod()` / `getConstructor()` that doesn't rely on reflection. Instead, `getMethodMetadata()` returns a new type `MethodMetadata`. This interface has methods similar to `java.lang.reflect.Method`, except it is implemented by a class generated during build specifically for each shape of intercepted method/constructor. For methods with small number of parameters or annotations, which is the most common case, the generated class has unfolded fields for parameters/annotations, which allows faster access than a data structure that's used otherwise (array or `Map`).
2. If an interceptor doesn't use the reflective methods mentioned above, it may be annotated `@Reflectionless`. If all interceptors that apply to some method/constructor are reflectionless, the `InvocationContext` methods `getMethod()` / `getConstructor()` will always return `null` when that method/constructor is intercepted, and the method/constructor is not registered for reflection (unless it is registered elsewhere).

The first commit is the infrastructure work, while the second is an example usage.

I measured performance impact using the JMH benchmark from this branch: https://github.com/Ladicek/arc-benchmarks/commits/reflectionless-interception

I measured RSS impact using this one-off tool: https://github.com/Ladicek/arc-crazybeans

# Benchmarks

## Measuring

All measurements were done on my otherwise-idle desktop machine (running Ubuntu 22.04 with kernel Linux 5.15.0-48-generic; hardware-wise it's Ryzen 5950X with 64 GB of RAM) with the following tuning:

```bash
# select the "performance" CPU scaling governor
sudo cpupower frequency-set -g performance

# disable hyperthreading
echo off | sudo tee /sys/devices/system/cpu/smt/control

# disable turbo boost
echo 0 | sudo tee /sys/devices/system/cpu/cpufreq/boost
```

(I know more tuning could/should be done, but I'm no expert...)

## Performance

`main` branch (82700f1f75455655accb335df85ec8d9fc258807):

```
Benchmark                                                                                   Mode  Cnt     Score    Error  Units
MethodAccessingInterceptionBenchmark.baseline_noInterceptor                                thrpt   25  4705.716 ±  1.563  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectiveWithAnnotationProceed       thrpt   25   551.225 ±  6.987  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectiveWithAnnotationShortCircuit  thrpt   25   579.019 ±  1.642  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectiveWithoutAnnotation           thrpt   25  1507.696 ± 17.745  ops/s
MethodAccessingInterceptionBenchmark.reflectiveWithAnnotationProceed                       thrpt   25   555.301 ±  1.868  ops/s
MethodAccessingInterceptionBenchmark.reflectiveWithAnnotationShortCircuit                  thrpt   25   580.690 ±  0.990  ops/s
MethodAccessingInterceptionBenchmark.reflectiveWithoutAnnotation                           thrpt   25  1519.545 ±  3.584  ops/s
```

this branch (4901aa9fb0ff57f3d30a6a0d085856688b340bf2):

```
Benchmark                                                                                       Mode  Cnt     Score     Error  Units
MethodAccessingInterceptionBenchmark.baseline_noInterceptor                                    thrpt   25  4707.689 ±   7.359  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectionlessWithAnnotationProceed       thrpt   25   841.232 ±  74.327  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectionlessWithAnnotationShortCircuit  thrpt   25   701.871 ± 105.664  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectionlessWithoutAnnotation           thrpt   25  1187.663 ± 127.325  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectiveWithAnnotationProceed           thrpt   25   460.182 ±  42.987  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectiveWithAnnotationShortCircuit      thrpt   25   537.209 ±  14.170  ops/s
MethodAccessingInterceptionBenchmark.manyAnnotations_reflectiveWithoutAnnotation               thrpt   25  1466.451 ±  15.869  ops/s
MethodAccessingInterceptionBenchmark.reflectionlessWithAnnotationProceed                       thrpt   25  1474.928 ±   7.062  ops/s
MethodAccessingInterceptionBenchmark.reflectionlessWithAnnotationShortCircuit                  thrpt   25  1307.191 ±   5.210  ops/s
MethodAccessingInterceptionBenchmark.reflectionlessWithoutAnnotation                           thrpt   25  1684.818 ±   8.631  ops/s
MethodAccessingInterceptionBenchmark.reflectiveWithAnnotationProceed                           thrpt   25   492.082 ±  47.662  ops/s
MethodAccessingInterceptionBenchmark.reflectiveWithAnnotationShortCircuit                      thrpt   25   493.029 ±  66.189  ops/s
MethodAccessingInterceptionBenchmark.reflectiveWithoutAnnotation                               thrpt   25  1467.469 ±  17.414  ops/s
```

## JVM RSS

`main` branch (reflection only): 106114.640 ± 580.154 kB
this branch (with `@Reflectionless`, so `getMethod()` returns `null`): 106108.000 ± 557.197 kB
this branch (without `@Reflectionless`, both methods are usable): 101162.720 ± 537.588 kB

The last line is pretty weird, but I ran the measurements several times and it's always the same. I don't have an explanation. Perhaps my RSS measurement methodology is wrong.

## Native RSS (GraalVM 22.2.0 Java 11 CE)

`main` branch (reflection only): 34404.400 ± 83.102 kB
this branch (with `@Reflectionless`, so `getMethod()` returns `null`): 33771.400 ± 75.389 kB
this branch (without `@Reflectionless`, both methods are usable): 35129.720 ± 80.170 kB

## Native binary size (GraalVM 22.2.0 Java 11 CE)

`main` branch (reflection only): 36226480 B
this branch (with `@Reflectionless`, so `getMethod()` returns `null`): 37500336 B
this branch (without `@Reflectionless`, both methods are usable): 37918128 B

# Conclusion

I originally generated 1 `MethodMetadata` implementation for each intercepted method. This was basically a bag of constants, so performance was even better than what's shown above. I later found that this has a huge impact on JVM RSS (in the same settings as benchmarks above, JVM RSS was 40 or 50 MB higher with this change compared to `main`). Interestingly, native RSS seemingly wasn't impacted at all.

I then refactored the code to the present approach. Performance still improves considerably, as shown above.

However, the RSS and native binary size impact is minuscule, which is the exact opposite of what I hoped for. My personal conclusion is that while I learned a lot, it was a waste of time otherwise.

I'm submitting this PR, as a draft, for your own consideration, but since this incurs significant maintenance cost for negligible gain, I propose we just don't merge it.