@mkouba 2022-10-25
I like the idea and even the numbers don't look bad at all!

However, I wonder if we could only generate the `MethodMetadata` for a `@Reflectionless` chain and use a reflection-based impl otherwise. This way an interceptor could use the new API but if a non-`@Reflectionless` interceptor is also used then just fallback to reflection and save some memory/native image footprint.