@Ladicek 2022-10-25
That is a good idea, and implementing `MethodMetadata` over `java.lang.reflect.Method` would be super-straightforward.