@Ladicek 2022-10-25
Good point re the missing test.

We could do whatever we want in `getMethod()` of a reflectionless intercepted method, we could possibly even throw an exception.