@manovotn 2022-10-25
> We could do whatever we want in `getMethod()` of a reflectionless intercepted method, we could possibly even throw an exception.

Yes, thinking about it more, throwing an exception makes more sense since once you mark the interceptor as reflectionless, it is an error to assume reflection access would work (even if the interceptor chain might then need it which is something you don't necessarily know at the time of writing)