@Ladicek 2022-10-25
That would be a little problematic, because the `InvocationContext` is not bound to an interceptor, but to the intercepted method.

I'll look at all the suggestions here and see what impact they make, but it will take a while. I want to get some actual productive work done first :-)