@manovotn 2022-10-25
> That would be a little problematic, because the InvocationContext is not bound to an interceptor, but to the intercepted method.

Fair enough, feel free to drop that suggestion, it was just a minor user-friendliness comment anyway :)