@Ladicek 2022-10-27
Closing this PR for now. I intend to experiment with the suggestions above, but at the moment, I don't believe the concept has enough of a value.